use new constructors from opslevel-go
